requirements.txt
numpy
plotly
pandas
python-mnist